jwt-python/requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
docopt==0.6.2
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.0
requests==2.22.0
six==1.12.0
urllib3==1.25.5
wcwidth==0.1.7
zipp==0.6.0
